norm対応 var - Makefileのnormをlibft, include, srcを対象にしました。
