+++ include/variable/env.h
@@ -6,7 +6,7 @@
 /*   By: ootsuboyoshiyuki <ootsuboyoshiyuki@stud    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/14 16:21:32 by ootsuboyosh       #+#    #+#             */
-/*   Updated: 2024/03/30 11:38:15 by ootsuboyosh      ###   ########.fr       */
+/*   Updated: 2024/04/11 18:18:08 by ootsuboyosh      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,11 +21,11 @@ void	set_envp(t_minishell *minish, const char **envp);
 char	**get_envp(t_minishell *minish);
 char	*get_var(t_minishell *minish, const char *key);
 void	add_or_update_var(t_minishell *minish, const char *key, const char *val,
-			e_var_type type);
+			t_var_type type);
 void	del_var(t_minishell *minish, const char *key);
 void	free_var(t_minishell *minish);
 bool	has_key(t_minishell *minish, const char *key);
-void	set_type(t_minishell *minish, const char *key, e_var_type type);
+void	set_type(t_minishell *minish, const char *key, t_var_type type);
 void	sort_key_list(char **key_list);
 char	**get_key_list(t_minishell *minish);
